Fix unused variable
