+++ tests/machine/x/pl/README.md
@@ -3,27 +3,10 @@
 This directory holds Prolog code generated from Mochi programs by the compiler tests.
 Below is a checklist of programs that were compiled and executed successfully.
 
-**19/21 programs compiled**
+**5/5 programs compiled**
 
-- [x] avg_builtin.mochi
-- [x] basic_compare.mochi
-- [x] binary_precedence.mochi
-- [x] bool_chain.mochi
-- [x] break_continue.mochi
-- [x] cast_string_to_int.mochi
-- [x] cast_struct.mochi
-- [x] closure.mochi
-- [x] count_builtin.mochi
-- [x] cross_join.mochi
-- [x] cross_join_filter.mochi
-- [x] cross_join_triple.mochi
-- [x] dataset_sort_take_limit.mochi
-- [x] dataset_where_filter.mochi
-- [x] for_list_collection.mochi
-- [x] for_loop.mochi
-- [x] let_and_print.mochi
 - [x] print_hello.mochi
+- [x] let_and_print.mochi
 - [x] var_assignment.mochi
-- [ ] append_builtin.mochi
-- [ ] exists_builtin.mochi
-
+- [x] for_loop.mochi
+- [x] for_list_collection.mochi
